FORCING - REFACTOR - Move forcing data I/O to ForcingTerms module
Refactors forcing data loading functionality to improve code organization
and separation of concerns.

Changes to src/ForcingTerms/ForcingTerms.jl:
- Added ForcingMode struct (moved from PerturbedEquilibrium)
- Added load_forcing_data!() function with generic signature
- Added load_forcing_ascii!() helper function
- Added load_forcing_hdf5!() helper function
- Exports ForcingMode and load_forcing_data!
- Imports DelimitedFiles and HDF5 for file I/O

Changes to src/PerturbedEquilibrium/:
- PerturbedEquilibrium.jl: Import ForcingMode and load_forcing_data! from ForcingTerms
- PerturbedEquilibrium.jl: Updated call site to use new function signature
- PerturbedEquilibriumStructs.jl: Removed ForcingMode (now in ForcingTerms)
- Response.jl: Removed load_forcing_data!() and helper functions

Benefits:
- ForcingTerms module now handles all forcing data I/O
- Cleaner separation between forcing data (ForcingTerms) and response calculations (PerturbedEquilibrium)
- More generic function signature allows reuse outside PerturbedEquilibrium
- Reduces coupling between modules
